ci(all): fix missing esm build files
Because npm pack ignores folders and files listed in gitignore,
the esm build directory for the `all` package was not included in the
tarball pushed to the registry.
I moved the gitignore line to the root of the repo and now it should
work properly. Letting us finally use the toolkit with vite for example!