refactor(action,button): remove hocs
The HOCs not only disturb the bundler's process to understand which code
does what, it also makes typing a pain.
Moreover, the HOC used here removed the react event and replaced it
with only the id of the HTML element.
This version makes typing slightly simpler, and gives user the complete
event while maintaining existing behaviour.